refactor: 쿠폰 페이지 버튼 가독성 좋은 이름으로 수정
